Center align the text inside the stat boxes
The HTML content such as title, value and actions in the stats boxes were left align. This commit changes the text alignment to center to improve the visual presentation and readability of these stats boxes. The text within the boxes was difficult to read and did not vertically align well with the icons, making it harder for users to quickly comprehend the information, especially on smaller screens. This change should make the pages on the website look more uniform and clean.